fix warnings, prepare for kuni-private external stuff
