+++ CMakeLists.txt
@@ -38,7 +38,7 @@ auib_mark_var_forwardable(AUI_COROUTINES)
 
 # import AUI
 auib_import(aui https://github.com/aui-framework/aui
-        COMPONENTS core json curl crypt image views audio
+        COMPONENTS core json curl crypt image views
         VERSION ${AUI_VERSION})
 
 
@@ -75,7 +75,6 @@ aui_link(${PROJECT_NAME} PUBLIC
         aui::crypt
         aui::image
         aui::views
-        aui::audio
         Td::TdStatic
         toml11::toml11
         prometheus-cpp::pull
@@ -85,13 +84,6 @@ if (APPLE)
     target_link_libraries(${PROJECT_NAME} PUBLIC "-framework CoreServices")
 endif()
 
-
-# Setup icon, display name, etc
-aui_app(TARGET ${PROJECT_NAME}
-        NAME "Kuni"
-        VENDOR "Alex2772"
-        ICON "assets/img/icon.svg"
-)
 
 # version info
 find_package(Git QUIET)
@@ -131,6 +123,17 @@ file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/version_info.cpp
 )
 target_sources(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_BINARY_DIR}/version_info.cpp)
 
+aui_compile_assets(${PROJECT_NAME})
+aui_enable_tests(${PROJECT_NAME})
+
+# Setup icon, display name, etc
+aui_app(TARGET ${PROJECT_NAME}
+        NAME "Kuni"
+        VENDOR "Alex2772"
+        ICON "assets/img/icon.svg"
+)
+
+
 # optional features
 foreach (_feature_switch KUNI_USE_FFMPEG)
     if (${${_feature_switch}})
@@ -154,9 +157,6 @@ if (KUNI_USE_FFMPEG)
     target_include_directories(Tests PRIVATE ${FFMPEG_INCLUDE_DIRS})
     target_link_libraries(Tests PUBLIC ${FFMPEG_LINK_LIBRARIES})
 endif ()
-
-aui_compile_assets(${PROJECT_NAME})
-aui_enable_tests(${PROJECT_NAME})
 
 if (EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/../kuni-private/CMakeLists.txt")
     # for developing private/instance-specific stuff
