+++ src/App.h
@@ -1,0 +1,751 @@
+#pragma once
+
+#include <random>
+#include <range/v3/action/insert.hpp>
+#include <range/v3/algorithm/any_of.hpp>
+#include <range/v3/algorithm/max_element.hpp>
+#include <range/v3/algorithm/min_element.hpp>
+#include <range/v3/numeric/accumulate.hpp>
+#include <range/v3/range/conversion.hpp>
+#include <range/v3/view/reverse.hpp>
+#include <range/v3/view/transform.hpp>
+
+#include "AUI/Common/AByteBuffer.h"
+#include "AUI/IO/AFileInputStream.h"
+#include "AUI/Curl/ACurl.h"
+#include "AUI/IO/APath.h"
+#include "AUI/Platform/Entry.h"
+#include "AUI/Util/ASharedRaiiHelper.h"
+#include "AUI/Util/kAUI.h"
+#include "AppBase.h"
+#include "IChatHistoryMessageProcessor.h"
+#include "ImageGenerator.h"
+#include "AUI/Image/jpg/JpgImageLoader.h"
+#include "telegram/ITelegramClient.h"
+#include "telegram/TelegramClientImpl.h"
+#include "StableDiffusionClientImpl.h"
+#include "OpenAIChatImpl.h"
+#include "OpenAIChatMeasurable.h"
+#include "Prometheus.h"
+#include "prompts.h"
+#include "ChatDatabase.h"
+#include "AUI/AppInfo.h"
+#include "llmui/image.h"
+#include "llmui/malicious_payloads.h"
+#include "llmui/telegram.h"
+#include "proxy_server/context_bridge.h"
+#include "tools/get_chat_photo.h"
+#include "tools/take_photo.h"
+#include "tools/record_audio.h"
+#include "tools/get_telegram_chats.h"
+#include "tools/react_with_emoji.h"
+#include "tools/search_chats.h"
+#include "tools/search_messages.h"
+#include "tools/view_messages_around.h"
+#include "tools/remove_and_ban_chat.h"
+#include "tools/leave_chat.h"
+#include "tools/join_chat.h"
+#include "tools/stickers.h"
+#include "tools/send_telegram_message.h"
+#include "tools/edit_message_text.h"
+#include "tools/forward_message.h"
+#include "ui/debug/KuniDebugWindow.h"
+#include "util/is_accessible_from_lockdown.h"
+#include "util/post_message.h"
+
+#include <range/v3/action/reverse.hpp>
+#include <range/v3/algorithm/contains.hpp>
+#include <range/v3/algorithm/count_if.hpp>
+#include <range/v3/algorithm/remove_if.hpp>
+#include <range/v3/algorithm/sort.hpp>
+
+#include "util/json_utils.h"
+
+#include <range/v3/view/take.hpp>
+#include <proxy_server/proxy_server.h>
+#include "tools/ask.h"
+#include "tools/group_admin_ban_user.h"
+#include "tools/group_admin_remove_message.h"
+#include "tools/group_admin_set_user_tag.h"
+#include "tools/remove_message.h"
+#include "tools/search_photo_in_gallery.h"
+
+
+#include <Diary.h>
+
+extern "C" AStringView project_version_info();
+extern std::default_random_engine gRandomEngine;
+
+class App : public AppBase {
+private:
+    static constexpr auto LOG_TAG = "App";
+
+public:
+    AVector<_<IChatHistoryMessageProcessor>> chatHistoryMessageProcessors;
+
+    App(_<ITelegramClient> telegram, _<IOpenAIChat> openAI)
+      : AppBase({ .workingDir = "data", .openAI = std::move(openAI) }), mTelegram(std::move(telegram)),
+        mChatDatabase(mTelegram)
+    {
+        ALOG_TRACE(LOG_TAG) << "App::App";
+        connect(mTelegram->onEvent, [this](AArc<const td::td_api::Object> event) {
+            td::td_api::downcast_call(const_cast<td::td_api::Object&>(*event), [&](const auto& u) {
+                mAsync << this->handleTelegramEvent(aui::ptr::alias(event, u));
+            });
+        });
+    }
+
+    [[nodiscard]] _<ITelegramClient> telegram() const { return mTelegram; }
+
+    void onOffline() override {
+        mCurrentlyOpenedChat.reset();
+        setOnline(false);
+    }
+
+    void onResponseAssembling(IOpenAIChat::Response response) override {
+        if (!mCurrentlyOpenedChat) {
+            return;
+        }
+
+
+        static std::chrono::high_resolution_clock::time_point lastEvent;
+        const auto now = std::chrono::high_resolution_clock::now();
+        using namespace std::chrono_literals;
+        if (now - lastEvent < 1s) {
+            // no need to spam.
+            return;
+        }
+        lastEvent = now;
+
+        if (!response.choices.empty()) {
+            const auto& choice = response.choices.at(0);
+            const auto& toolCalls = choice.message.tool_calls;
+            for (const auto& tc : toolCalls) {
+                if (tc.function.name == "sticker_send") {
+                    mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::sendChatAction(
+                        mCurrentlyOpenedChat->chat->id_,
+                        {},
+                        {},
+                        ITelegramClient::toPtr(td::td_api::chatActionChoosingSticker()))));
+                    break;
+                }
+                if (tc.function.name == "record_audio") {
+                    mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::sendChatAction(
+                        mCurrentlyOpenedChat->chat->id_,
+                        {},
+                        {},
+                        ITelegramClient::toPtr(td::td_api::chatActionRecordingVoiceNote()))));
+                    break;
+                }
+                if (tc.function.name == "take_photo") {
+                    mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::sendChatAction(
+                        mCurrentlyOpenedChat->chat->id_,
+                        {},
+                        {},
+                        ITelegramClient::toPtr(td::td_api::chatActionUploadingPhoto()))));
+                    break;
+                }
+            }
+        }
+        mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::sendChatAction(
+            mCurrentlyOpenedChat->chat->id_, {}, {}, ITelegramClient::toPtr(td::td_api::chatActionTyping()))));
+    }
+
+    void updateTools(OpenAITools& actions, const IOpenAIChat::Session& temporaryContext) override {
+        AppBase::updateTools(actions, temporaryContext);
+        if (config().capabilityTakePhoto) {
+            actions.insert(tools::takePhoto(_new<StableDiffusionClientImpl>(), openAI()));
+        }
+        actions.insert(tools::searchPhotoInGallery(openAI(), temporaryContext));
+        if (config().capabilityRecordVoice) {
+            actions.insert(tools::recordAudio());
+        }
+        actions.insert(tools::getTelegramChats(telegram(), openAI(), isActingProactively()));
+        actions.insert(tools::searchChats(telegram()));
+        actions.insert(tools::searchMessages(telegram(), openAI(), temporaryContext));
+        actions.insert(tools::viewMessagesAround(telegram(), openAI(), temporaryContext));
+        actions.insert(tools::removeAndBanChat(telegram()));
+        actions.insert({
+                .name = "open_chat_by_id",
+                .description = "Opens a chat by its id. Use this to start conversation. Use get_telegram_chats to "
+                               "retrieve `chat_id`s.",
+                .parameters =
+                    {
+                        .properties =
+                            {
+                                {"chat_id", {.type = "integer", .description = "The ID of the Telegram chat"}},
+                            },
+                        .required = {"chat_id"},
+                    },
+                .handler = [this](OpenAITools::Ctx ctx) -> AFuture<AString> {
+                    if (ranges::count_if(ctx.allToolCalls, [](const IOpenAIChat::Message::ToolCall& call) {
+                        return call.function.name == "open_chat_by_id";
+                    }) > 1) {
+                        co_return "You can only call this tool once per turn.";
+                    }
+
+                    auto chatId = util::jsonAsLongInt(ctx.args["chat_id"]).valueOrException("chat_id integer is required");
+
+                    // Check lockdown mode - only allow PAPIK_CHAT_ID if lockdown is enabled
+
+                    if (! co_await util::isAccessibleFromLockdown(*telegram(), chatId)) {
+                        ALogger::err(LOG_TAG) << "Error: Lockdown mode is enabled. You can only open chat with ID {} (PAPIK_CHAT_ID)."_format(config().papikChatId);
+                        co_return "No such chat";
+                    }
+
+                    co_return co_await llmuiOpenTelegramChat(ctx.logger, ctx.tools, chatId, ctx.temporaryContext);
+                },
+            });
+        if (config().canJoinChats) {
+            actions.insert({
+                    .name = "join_chat_by_link",
+                    .description = "Joins a chat/channel by its invite link (e.g. "
+                                   "https://t.me/joinchat/xxxx) and opens it immediately, just like tapping the "
+                                   "invite link in the official Telegram client.",
+                    .parameters =
+                        {
+                            .properties =
+                                {
+                                    {"invite_link", {.type = "string", .description = "The invite link, e.g. https://t.me/joinchat/xxxx"}},
+                                },
+                            .required = {"invite_link"},
+                        },
+                    .handler = [this](OpenAITools::Ctx ctx) -> AFuture<AString> {
+                        auto inviteLink = ctx.args["invite_link"].asStringOpt().valueOrException("invite_link string is required");
+
+                        int64_t chatId;
+                        try {
+                            auto joinedChat = co_await telegram()->sendQueryWithResult(
+                                ITelegramClient::toPtr(td::td_api::joinChatByInviteLink(inviteLink.toStdString())));
+                            chatId = joinedChat->id_;
+                        } catch (const AException& e) {
+                            ALogger::err(LOG_TAG) << "Failed to join chat by invite link \"" << inviteLink << "\": " << e;
+                            co_return "Error: failed to join chat by invite link: {}"_format(e.getMessage());
+                        }
+
+                        co_return co_await llmuiOpenTelegramChat(ctx.logger, ctx.tools, chatId, ctx.temporaryContext);
+                    },
+                });
+        }
+        if (config().capabilityUseStickers) {
+            actions.insert(tools::stickers::list(telegram(), openAI()));
+            actions.insert(tools::stickers::save(telegram()));
+        }
+    }
+
+    AFuture<> sendNotificationsOnInit() {
+        if (!config().checkChatsOnStartup) {
+            co_return;
+        }
+        // tdlib does not send notifications for unread chats on program startup. we'll fix this.
+        auto chats = co_await getChats();
+        chats |= ranges::actions::reverse;   // older first, newest last
+        for (auto& chat : chats) {
+            if (chat->unread_count_ == 0) {
+                continue;
+            }
+            // make up a updateNewMessage event and pass it to handleTelegramEvent. the latter will format a
+            // notification for us.
+            auto notification = _new<td::td_api::updateNewMessage>();
+            notification->message_ = std::move(chat->last_message_);
+            co_await handleTelegramEvent(std::move(notification));
+        }
+    }
+
+
+protected:
+    AString onCleanContext() const override {
+        AString result = AppBase::onCleanContext();
+        // Alex2772 (9 Jul 2026):
+        // consumes too much content.
+        // if llm wants to share a sticker, it would call get_stickers().
+        //
+        // if (config().capabilityUseStickers) {
+        //     auto list = co_await llmui::listFavoriteStickers(*telegram(), *openAI());
+        //     if (!list.empty()) {
+        //         result += "<your_favorite_stickers>\n";
+        //         result += list;
+        //         result += "</your_favorite_stickers>\n";
+        //     }
+        // }
+        return result;
+    }
+
+private:
+    struct CurrentlyOpenedChat {
+        App& app;
+        _<td::td_api::chat> chat;
+
+        ~CurrentlyOpenedChat() {
+            app.mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::sendChatAction(chat->id_, {}, {}, nullptr)));
+            app.mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::closeChat(chat->id_)));
+        }
+    };
+
+    _<ITelegramClient> mTelegram;
+    std::list<MetricsBreadcumbs::Point> mLastOpenedChatLastMetrics;
+    AOptional<CurrentlyOpenedChat> mCurrentlyOpenedChat;
+    ChatDatabase mChatDatabase;
+
+    AMap<AString /* path */, AString /* description */> mImages = {};
+
+    AFuture<AVector<_<td::td_api::chat>>> chatIdsToChats(std::span<td::td_api::int53> ids) {
+        auto chats = ids | ranges::view::transform([&](td::td_api::int53 chatId) { return telegram()->getChat(chatId); }) | ranges::to_vector;
+        AVector<_<td::td_api::chat>> result;
+        result.reserve(chats.size());
+        for (const auto& chat : chats) {
+            result.push_back(co_await chat);
+        }
+        co_return result;
+    }
+
+    AFuture<_<td::td_api::chat>> chatIdToChat(td::td_api::int53 id) { co_return co_await telegram()->getChat(id); }
+
+    AFuture<AVector<_<td::td_api::chat>>> getChats() {
+        auto chatList = co_await telegram()->sendQueryWithResult(
+            ITelegramClient::toPtr(td::td_api::getChats(ITelegramClient::toPtr(td::td_api::chatListMain()), 200)));
+        co_return co_await chatIdsToChats(chatList->chat_ids_);
+    }
+
+    template <aui::derived_from<td::td_api::Object> Object>
+    AFuture<> handleTelegramEvent(_<const Object> u) {
+        TelegramClientImpl::StubHandler {}(*u);
+        co_return;
+    }
+
+    AFuture<AOptional<AString /* msg */>> tryHandleCmd(int64_t senderId, AStringView msg) {
+        try {
+            if (msg == "/version") {
+                static constexpr char KERNEL_NAME[] = {
+                    'k', 'u', 'n', 'i', 0   // original kernel name, plz do not replace
+                };
+#if AUI_TESTS_MODULE
+                co_return "Kernel: {}"_format(KERNEL_NAME);
+#else
+                co_return "{}\nKernel: {}"_format(project_version_info(), KERNEL_NAME);
+#endif
+            }
+        } catch (const AException& e) {
+            ALogger::err(LOG_TAG) << "Failed to handle command: " << e;
+        }
+        co_return std::nullopt;
+    }
+
+    AFuture<> handleTelegramEvent(AArc<const td::td_api::updateNewMessage> u) {
+        int64_t userId = 0;
+        if (auto user = ITelegramClient::tryCastTo<const td::td_api::messageSenderUser>(*u->message_->sender_id_)) {
+            userId = user->user_id_;
+        }
+        if (userId == mTelegram->myId()) {
+            co_return;
+        }
+
+        // Check lockdown mode - only allow PAPIK_CHAT_ID if lockdown is enabled
+        if (!co_await util::isAccessibleFromLockdown(*telegram(), u->message_->chat_id_)) {
+            co_return;
+        }
+        if (!co_await util::isAccessibleFromLockdown(*telegram(), u->message_->chat_id_, config().chatNotificationFilter)) {
+            co_return;
+        }
+
+        auto chat = co_await mTelegram->getChat(u->message_->chat_id_);
+
+        if (chat->notification_settings_) {
+            if (chat->notification_settings_->mute_for_ > 0) {
+                // Alex2772 (Apr 23 2026):
+                //
+                // Added a probability to ignore a muted chat.
+                //
+                // If we always ignore a muted chat, i.e.,
+                // ```cpp
+                // co_return;
+                // ```
+                // LLM will read this only if:
+                // - it occasionally called get_telegram_chats, and
+                // - it recognized a telegram chat with a lot of messages, and
+                // - it decided to read it
+                // which basically means LLM will NEVER read a muted chat.
+                //
+                // I've added a PROBABILITY to ignore a muted chat. This allows the account holder to mute the chat,
+                // so LLM will give a lot less attention to it. This is useful for spammy chat.
+                // If the account holder wants Kuni to ignore the chat completely, they should archive the chat.
+                if (std::uniform_real_distribution<>(0.0, 1.0)(gRandomEngine) < 0.8) {
+                    co_return;
+                }
+            }
+        }
+        auto notification = "<notification chat_id=\"{}\">\n"_format(chat->id_);
+
+        if (userId == u->message_->chat_id_) {
+            if (auto cmdResponse = co_await tryHandleCmd(
+                    userId, llmui::extractMessageTypeAndText(const_cast<td::td_api::message&>(*u->message_)))) {
+                co_await util::telegramPostMessage(
+                    *telegram(), userId, std::move(*cmdResponse), std::nullopt, std::nullopt, u->message_->id_);
+                co_return;
+            }
+            notification += "You received a direct message from {} (chat_id = {})"_format(chat->title_, chat->id_);
+        } else if (userId != 0) {
+            auto user = co_await mTelegram->getUser(userId);
+            notification += "{} {} (user_id = {}) sent a message in group chat \"{}\" (chat_id = {})"_format(
+                user->first_name_, user->last_name_, userId, chat->title_, chat->id_);
+        } else {
+            notification += "Channel \"{}\" (chat_id={}) created a new post\n"_format(chat->title_, chat->id_);
+        }
+        notification +=
+            "\n</notification>\n"
+            "You don't have any chat open. Use #open tool to open the chat";
+
+        const int priority = [&] {
+            if (userId == config().papikChatId) {
+                return 1000;
+            }
+            if (config().wakeUpOnPinnedChat) {
+                for (const auto& position : chat->positions_) {
+                    if (position->is_pinned_) {
+                        return 100;
+                    }
+                }
+            }
+            return 0;
+        }();
+
+        notificationManager().passNotificationToAI(NotificationManager::Notification{
+            .message = std::move(notification),
+            .actions = {
+              {
+                .name = "open",
+                .description = "Open \"{}\" chat. Use this if you'd like to reply or see messages."_format(chat->title_),
+                .handler = [this, chatId = chat->id_](OpenAITools::Ctx ctx) -> AFuture<AString> {
+                    return llmuiOpenTelegramChat(ctx.logger, ctx.tools, chatId, ctx.temporaryContext);
+                },
+              },
+            },
+            .priority = priority,
+            .pin = "<chat id=\"{}\" />"_format(chat->id_),
+        });
+
+        if (priority > 0) {
+            wakeUpIfSleeping();
+        }
+
+        co_return;
+    }
+
+    void setOnline(bool online = true) {
+        mTelegram->sendQuery(ITelegramClient::toPtr(
+            td::td_api::setOption("online", ITelegramClient::toPtr(td::td_api::optionValueBoolean(online)))));
+    }
+
+
+public:
+    AFuture<AString> llmuiOpenTelegramChat(ALogger& logger, OpenAITools& tools, int64_t chatId, const IOpenAIChat::Session& temporaryContext) {
+        // Check lockdown mode - only allow PAPIK_CHAT_ID if lockdown is enabled
+        if (!co_await util::isAccessibleFromLockdown(*telegram(), chatId)) {
+            logger.err(LOG_TAG) << "Error: Lockdown mode is enabled. You can only open chat with ID {} (PAPIK_CHAT_ID)."_format(
+                config().papikChatId);
+            co_return "No such chat";
+        }
+
+        co_await telegram()->waitForConnection();
+        setOnline();
+        mTelegram->sendQuery(ITelegramClient::toPtr(td::td_api::openChat(chatId)));
+        notificationManager().removeNotifications("<notification chat_id=\"{}\">\n"_format(chatId));
+
+        _<td::td_api::chat> chat = co_await mTelegram->getChat(chatId);
+        mCurrentlyOpenedChat.emplace(*this, chat);
+        mLastOpenedChatLastMetrics = std::list<MetricsBreadcumbs::Point> {};
+        mLastOpenedChatLastMetrics.emplace_back(metricBreadcumbs(), "chat", chat->title_);
+
+        AString result;
+
+        // loaded messages. first goes the newest, last goes the oldest
+        td::td_api::array<td::td_api::object_ptr<td::td_api::message>> messages;
+        co_await [&]() -> AFuture<> {
+            int64_t fromMessage = 0;
+            for (;;) {
+                auto response = co_await mTelegram->sendQueryWithResult(
+                    ITelegramClient::toPtr(td::td_api::getChatHistory(chatId, fromMessage, 0, 30, false)));
+                if (response->messages_.empty()) {
+                    break;
+                }
+                fromMessage = response->messages_.back()->id_;
+                size_t length = 0;
+                for (auto& msg : response->messages_) {
+#if AUI_DEBUG
+                    AUI_ASSERT(!ranges::any_of(messages, [&](const auto& m) { return m->id_ == msg->id_; }));
+#endif
+                    const auto msgFormatting = R"(<message message_id="{}")"_format(msg->id_);
+                    length += to_string(msg->content_).length();
+                    messages.push_back(std::move(msg));
+                    if (ranges::any_of(temporaryContext, [&](const IOpenAIChat::Message& msg) {
+                            return msg.content.contains(msgFormatting);
+                        })) {
+                        // this message is already in context, which means we don't need to load further.
+                        // we'll just reassure this one, so the continuation of a dialogue in context won't feel
+                        // detached, and stop at this point.
+                        co_return;
+                    }
+                    if (messages.size() < 3) {
+                        continue;
+                    }
+                    if (length >= config().chatMaxHistoryLength) {
+                        co_return;
+                    }
+                }
+            }
+        }();
+        logger.info(LOG_TAG) << "Loaded " << messages.size() << " message(s): " << chat->title_;
+
+        // Compute response-time metadata for Prometheus. messages[0] is the most recent.
+        mLastOpenedChatLastMessageTime = [&]() -> AOptional<std::chrono::system_clock::time_point> {
+            if (messages.empty()) {
+                return std::nullopt;
+            }
+            return std::chrono::system_clock::from_time_t(messages.front()->date_);
+        }();
+        mLastOpenedChatLastMetrics.emplace_back(metricBreadcumbs(), "scenario", [&] {
+            if (messages.empty()) {
+                // this means Kuni sent a message to a new person.
+                return "new conversation";
+            }
+
+            td::td_api::int53 senderId = 0;
+            td::td_api::downcast_call(
+                *messages.front()->sender_id_,
+                aui::lambda_overloaded {
+                  [&](td::td_api::messageSenderUser& user) { senderId = user.user_id_; },
+                  [](auto&) {},
+                });
+            if (senderId == mTelegram->myId()) {
+                // last message is from Kuni.
+                return "kuni proactive";
+            }
+            return "reply to user";
+        }());
+        if (messages.empty()) {
+            // Kuni sometimes opens random chats?
+            // throw AException("Failed to open chat");
+
+            if (config().canWriteToANewPerson) {
+                result += "This chat is empty! Only proceed if you looked up a @username and it led you here.\n";
+                result +=
+                    "Only write what you have to say to the chat; if someone asked you to text this person, just text "
+                    "them.\n";
+                result +=
+                    "If you try to get back to the original chat and type something, you will be sending an extra "
+                    "message to the wrong chat.";
+            }
+            // goto naxyi;
+        }
+        bool isAdmin = false;
+        {
+            for (auto& msg : messages | ranges::view::reverse) {
+                auto msgFormatted =
+                    co_await llmui::formatChatHistoryMessage(*telegram(), *msg, *chat, *openAI(), temporaryContext);
+                for (const auto& i : chatHistoryMessageProcessors) {
+                    msgFormatted = co_await i->processChatHistoryMessage(*chat, *msg, std::move(msgFormatted));
+                }
+                result += msgFormatted;
+                td::td_api::int53 senderId = 0;
+                td::td_api::downcast_call(
+                    *msg->sender_id_,
+                    aui::lambda_overloaded {
+                      [&](td::td_api::messageSenderUser& user) { senderId = user.user_id_; },
+                      [](auto&) {},
+                    });
+                if (senderId == mTelegram->myId()) {
+                    td::td_api::downcast_call(
+                        *msg->content_,
+                        aui::lambda_overloaded {
+                          [&](td::td_api::messageText& text) {
+                              llmui::checkForMaliciousPayloads(text.text_->text_);
+                              if (text.link_preview_) {
+                                  result += "\n" + llmui::formatLinkPreview(*text.link_preview_);
+                              }
+                          },
+                          [](auto& i) {},
+                        });
+                } else {
+                    // store message with confidence=1 for future reference.
+                    // storing it with sender and message_id so LLM can refer to this message (i.e., forward it
+                    // or reply to it if contradictions was found)
+
+                    // not sure if this is needed; i think LLM would be confused if <message> tag exists in both
+                    // diary and current chat listing.
+                    //
+                    // currently disabled because it pollutes diary very quickly and according to kuni --debug,
+                    // its hard to find something meaningful; instead you get a bunch of messages
+                    //
+                    // auto msgReformatted = msgFormatted
+                    //     .replacedAll("<message", "<m")
+                    //     .replacedAll("</message", "</m")
+                    //     .replacedAll("unread", "")
+                    // ;
+                    // diary().save(Diary::EntryEx{
+                    //     .id = "msg_{}"_format(msg->id_),
+                    //     .metadata = {
+                    //         // confidence=1 means this is a fact and not LLM's AI slop.
+                    //         // sleep consolidator can't alter entries with confidence=1.
+                    //         .confidence = 1.f,
+                    //     },
+                    //     .freeformBody = std::move(msgReformatted),
+                    // });
+                }
+            }
+
+            if (!messages.empty()) {
+                mTelegram->sendQuery(
+                    ITelegramClient::toPtr(td::td_api::viewMessages(chatId, td::td_api::array<td::td_api::int53>{messages.front()->id_}, nullptr, true)));
+            }
+
+            auto prefix = "You switched to the chat \"{}\" in Telegram."_format(chat->title_);
+
+            {
+                const auto tag = "<previous_history chat_id=\"{}\">"_format(chatId);
+                if (!ranges::any_of(temporaryContext, [&](const IOpenAIChat::Message& ctx) {
+                    return ctx.content.contains(tag);
+                })) {
+                    if (auto lastAsk = mChatDatabase.getLastAskResult(chatId)) {
+                        prefix += "\n{}\n{}\n</previous_history>\n"_format(tag, *lastAsk);
+                    }
+                }
+            }
+
+            prefix += " You see last messages:\n";
+            result = prefix + result;
+
+            // Mirror the official Telegram client behavior: if you open a group chat/channel you haven't joined,
+            // you only see a "Join" button (plus the ability to react to messages) instead of a text field.
+            if (chat->type_->get_id() == td::td_api::chatTypeBasicGroup::ID ||
+                chat->type_->get_id() == td::td_api::chatTypeSupergroup::ID) {
+                bool isMember = true;
+                try {
+                    auto member = co_await mTelegram->sendQueryWithResult(ITelegramClient::toPtr(td::td_api::getChatMember(
+                        chatId, ITelegramClient::toPtr(td::td_api::messageSenderUser(mTelegram->myId())))));
+                    switch (member->status_->get_id()) {
+                        case td::td_api::chatMemberStatusLeft::ID:
+                        case td::td_api::chatMemberStatusBanned::ID:
+                            isMember = false;
+                            break;
+                        case td::td_api::chatMemberStatusCreator::ID:
+                        case td::td_api::chatMemberStatusAdministrator::ID:
+                            isAdmin = true;
+                            break;
+                        default:
+                            break;
+                    }
+                } catch (const AException& e) {
+                    // either:
+                    // USER_NOT_PARTICIPANT - treat as not a member.
+                    // or - member list is not accessible (for large channels)
+                    // fallback to chat_lists.
+                    isMember = !chat->chat_lists_.empty();
+                }
+
+                if (!isMember) {
+                    result += R"(
+<instructions>
+You are NOT a member of "{}". You can't send messages here; you can only #react_with_emoji to messages, or use
+#join_chat to become a member (after which you'll be able to fully participate the next time you open this chat).
+</instructions>
+)"_format(chat->title_);
+                    tools = OpenAITools {
+                        tools::reactWithEmoji(telegram(), chat),
+                    };
+                    if (config().canJoinChats) {
+                        tools.insert(tools::joinChat(telegram(), chat));
+                    }
+                    co_return result;
+                }
+            }
+
+            switch (chat->type_->get_id()) {
+                case td::td_api::chatTypeSecret::ID:
+                case td::td_api::chatTypePrivate::ID:
+                    result += fmt::format(
+                        R"(
+<instructions>
+You are in private chat with {} (also known as direct messages or DM).
+
+{}
+</instructions>
+)",
+                        chat->title_, prompts().messagesEpilogue);
+
+                    break;
+                case td::td_api::chatTypeBasicGroup::ID:
+                basicGroup:
+                    result += R"(
+<instructions>
+You are in group chat called \"{}\".
+
+{}
+</instructions>
+)"_format(chat->title_, prompts().messagesEpilogue);
+                    break;
+                case td::td_api::chatTypeSupergroup::ID: {
+                    if (!static_cast<td::td_api::chatTypeSupergroup&>(*chat->type_).is_channel_) {
+                        // lol what?
+                        goto basicGroup;
+                    }
+                    result += R"(
+<instructions>
+You are in telegram channel (also known as supergroup) called \"{}\".
+Pay close attention to these messages. Acquire context from them. You can't respond in telegram channels
+(#send_telegram_message tool is not available). Instead, do what you usually do when reading newsletters: reflect and reason
+on them.
+Some channels have reactions enabled. In that case, you can sometimes react with #react_with_emoji to express your feelings about a message, but you can't send a full reply.
+
+Forwarding posts:
+If you find a post genuinely interesting, funny, or relevant to someone you know — you can forward it to another chat
+using #forward_message. Be selective: only forward posts that are truly worth sharing. You can add a short comment
+expressing your reaction. Use #get_telegram_chats to find the destination chat_id if needed.
+Do NOT forward ads, sponsored posts, or low-value content.
+</instructions>
+)"_format(chat->title_);
+                    tools = OpenAITools {
+                        tools::reactWithEmoji(telegram(), chat),
+                        tools::forwardMessage(telegram(), chat),
+                    };
+                    co_return result;   // no send_telegram_message for channels
+                }
+            }
+        }
+
+    naxyi:
+        tools = OpenAITools {
+            tools::sendTelegramMessage(telegram(), openAI(), chat, _new<td::td_api::array<td::td_api::object_ptr<td::td_api::message>>>(std::move(messages))),
+            tools::getChatPhoto(telegram(), openAI(), chat, temporaryContext),
+            tools::reactWithEmoji(telegram(), chat),
+            tools::removeMessage(telegram(), chat),
+            tools::editMessageText(telegram(), chat),
+            tools::forwardMessage(telegram(), chat),
+        };
+
+        if (config().capabilityUseStickers) {
+            tools.insert(tools::stickers::send(telegram(), chat));
+        }
+
+        switch (chat->type_->get_id()) {
+            case td::td_api::chatTypeBasicGroup::ID:
+            case td::td_api::chatTypeSupergroup::ID:
+
+                if (config().canLeaveChats) {
+                    tools.insert(tools::leaveChat(telegram(), chat));
+                }
+                if (isAdmin) {
+                    tools.insert(tools::groupAdminRemoveMessage(telegram(), chat));
+                    tools.insert(tools::groupAdminBanUser(telegram(), chat));
+                    tools.insert(tools::groupAdminSetUserTag(telegram(), chat));
+                }
+                break;
+            default:
+                break;
+        }
+
+        tools.insert(toolAsk(temporaryContext));
+        mChatDatabase.patchAskTool(tools, chatId);
+
+        co_return result;
+    }
+};
